+++ src/domain/account/account.enum.ts
@@ -1,0 +1,3 @@
+export enum AccountCategory {
+  EMAIL = 'email',
+}
