+++ src/domain/account/dto/account.dto.ts
@@ -1,0 +1,21 @@
+import { AccountCategory } from '@domain/account/account.enum';
+
+export interface AccountDTO {
+  /**
+   * unique identification for accounts, such as email
+   * @type string
+   */
+  identification: string;
+
+  /**
+   * password for accounts, it should be one-way encrypted
+   * @type string
+   */
+  password: string;
+
+  /**
+   * category for accounts, now only support 'EMAIL'
+   * @type string
+   */
+  category: AccountCategory;
+}
