+++ src/domain/user/dto/user.dto.ts
@@ -1,9 +1,6 @@
 import typia from 'typia';
-import {
-  AccountCategory,
-  AgreementCategory,
-  AreaCategory,
-} from '@domain/user/user.enum';
+import { AgreementCategory, AreaCategory } from '@domain/user/user.enum';
+import { AccountDTO } from '@domain/account/dto/account.dto';
 
 export const checkRegisterAccountDto = typia.createIs<RegisterUserDTO>();
 
@@ -31,26 +28,6 @@ export interface RegisterUserDTO {
    * @type Array
    */
   agreement: AgreementDTO[];
-}
-
-export interface AccountDTO {
-  /**
-   * unique identification for accounts, such as email
-   * @type string
-   */
-  identification: string;
-
-  /**
-   * password for accounts, it should be one-way encrypted
-   * @type string
-   */
-  password: string;
-
-  /**
-   * category for accounts, now only support 'EMAIL'
-   * @type string
-   */
-  category: AccountCategory;
 }
 
 export interface AgreementDTO {
