+++ src/domain/user/service/user.service.ts
@@ -1,17 +1,16 @@
 import { Injectable } from '@nestjs/common';
 import { UserRepository } from '@domain/user/repository/user.repository';
-import { AccountRepository } from '@domain/user/repository/account.repository';
 import { EndUser } from '@domain/user/core/end-user';
-import { Account } from '@domain/user/core/account';
 import { RegisterUserDTO } from '@domain/user/dto/user.dto';
 import { AgreementRepository } from '@domain/user/repository/agreements.repository';
 import { AreaRepository } from '@domain/user/repository/area.repository';
+import { AccountService } from '@domain/user/service/account.service';
 
 @Injectable()
 export class UserService {
   constructor(
+    private accountService: AccountService,
     private userRepo: UserRepository,
-    private accountRepo: AccountRepository,
     private agreementRepo: AgreementRepository,
     private areaRepo: AreaRepository,
   ) {}
@@ -28,12 +27,7 @@ export class UserService {
       dto.userProperty,
     );
 
-    const account = new Account(this.accountRepo, {
-      identification: dto.account.identification,
-      password: dto.account.password,
-      category: dto.account.category,
-    });
-    await account.register(user.id);
+    await this.accountService.register(user.id, dto.account);
     return user;
   }
 }
