+++ src/domain/user/user.enum.ts
@@ -1,9 +1,6 @@
 export enum UserState {
   ACTIVE = 'active',
 }
-export enum AccountCategory {
-  EMAIL = 'email',
-}
 
 export enum AgreementCategory {
   PERSONAL_INFORMATION = 'personal_information',
